# api/gen/http/openapi.yaml
swagger: "2.0"
info:
    title: PgQueryNarrative API
    description: Data Storyteller AI - Convert PostgreSQL SQL queries to business narratives
    version: v1
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /api/v1/reports/generate:
        post:
            tags:
                - reports
            summary: generate reports
            description: Generate a narrative report from a SQL query
            operationId: reports#generate
            parameters:
                - name: GenerateRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/GenerateReportPayload'
                    required:
                        - sql
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/Report'
                        required:
                            - id
                            - sql
                            - narrative
                            - metrics
                            - created_at
                            - llm_model
                            - llm_provider
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/ValidationError'
                        required:
                            - name
                            - message
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/LLMError'
                        required:
                            - name
                            - message
            schemes:
                - http
definitions:
...
    ChartSuggestion:
        title: ChartSuggestion
        type: object
        properties:
            chart_type:
                type: string
                description: 'Chart type identifier: bar, line, pie, area, table'
                example: Sit culpa corrupti minus quo.
            label:
                type: string
                description: Human-readable label
                example: Enim et adipisci.
            reason:
                type: string
                description: Why this chart fits the data
                example: Qui maxime quidem velit laboriosam magni consectetur.
        example:
            chart_type: Beatae error dicta praesentium quia itaque.
            label: Voluptatem ad animi.
            reason: In fugiat possimus labore vero ut id.
        required:
            - chart_type
            - label
            - reason
...
    GenerateReportPayload:
        title: GenerateReportPayload
        type: object
        properties:
            saved_query_id:
                type: string
                example: 4da12188-18da-4261-a80d-752215d1fa2c
                format: uuid
            sql:
                type: string
                example: 24p
                pattern: ^[^;]+$
                minLength: 1
                maxLength: 10000
        example:
            saved_query_id: 460f4aee-4b19-4854-846c-1fc337e891d1
            sql: xf
        required:
            - sql
    LLMError:
        title: LLMError
        type: object
        properties:
            code:
                type: string
                example: Repellendus aut minus velit vel debitis.
            message:
                type: string
                example: Recusandae voluptatem in ea.
            name:
                type: string
                example: Optio reiciendis laboriosam veritatis consequatur in ipsam.
        example:
            code: Non labore eveniet nesciunt.
            message: Fugit voluptates nulla in doloribus sit tempora.
            name: Velit unde.
        required:
            - name
            - message
...
    NarrativeContent:
        title: NarrativeContent
        type: object
        properties:
            drivers:
                type: array
                items:
                    type: string
                    example: Illo quis fugit ipsum corrupti sapiente.
                example:
                    - Aut illum sunt sed et.
                    - Nihil quod a in et doloremque eius.
            headline:
                type: string
                example: Odio nisi et velit quia nesciunt id.
            limitations:
                type: array
                items:
                    type: string
                    example: Quod laudantium atque dolor harum facere deleniti.
                example:
                    - Quia et debitis.
                    - Est inventore autem non.
                    - Odio repudiandae.
                    - Commodi eum.
            recommendations:
                type: array
                items:
                    type: string
                    example: Totam est rerum totam beatae et quasi.
                example:
                    - Quia perferendis atque aut eos.
                    - Beatae et quo rem natus dolor aut.
            takeaways:
                type: array
                items:
                    type: string
                    example: Libero est rerum in.
                example:
                    - Molestiae optio ut eveniet maiores.
                    - Id voluptas ea occaecati quia cum.
        example:
            drivers:
                - Fugit id voluptatem rerum quas.
                - Nobis est voluptas.
                - Qui odio laudantium.
                - Ex quos ad dolor distinctio ea.
            headline: Earum modi aut rerum molestias velit.
            limitations:
                - Non blanditiis dolore rem quam.
                - Dolorem deleniti repudiandae ab aut et.
                - Facere consequuntur eaque provident.
            recommendations:
                - Et est harum at.
                - Aliquam et eligendi sunt dolorem recusandae.
            takeaways:
                - Omnis dolores ad inventore est dolorem sed.
                - Rerum et.
                - Ex nihil harum tenetur eaque.
        required:
            - headline
            - takeaways
...
    Report:
        title: Report
        type: object
        properties:
            chart_suggestions:
                type: array
                items:
                    $ref: '#/definitions/ChartSuggestion'
                description: Suggested chart types based on result shape
                example:
                    - chart_type: Vel sed delectus dolorem voluptatibus.
                      label: Culpa eveniet et.
                      reason: Ratione maxime nulla.
                    - chart_type: Vel sed delectus dolorem voluptatibus.
                      label: Culpa eveniet et.
                      reason: Ratione maxime nulla.
            created_at:
                type: string
                example: "1975-01-25T11:11:31Z"
                format: date-time
            id:
                type: string
                example: 76e4c40c-45de-44ed-8720-4cf6c70ec90d
                format: uuid
            llm_model:
                type: string
                example: Expedita cumque sunt non.
            llm_provider:
                type: string
                example: Animi veniam rem.
            metrics:
                $ref: '#/definitions/MetricsData'
            narrative:
                $ref: '#/definitions/NarrativeContent'
            saved_query_id:
                type: string
                example: fc051b9e-16a4-4017-9238-6a6b8fc0d67e
                format: uuid
            sql:
                type: string
                example: Harum molestiae perspiciatis.
        example:
            chart_suggestions:
                - chart_type: Vel sed delectus dolorem voluptatibus.
                  label: Culpa eveniet et.
                  reason: Ratione maxime nulla.
                - chart_type: Vel sed delectus dolorem voluptatibus.
                  label: Culpa eveniet et.
                  reason: Ratione maxime nulla.
                - chart_type: Vel sed delectus dolorem voluptatibus.
                  label: Culpa eveniet et.
                  reason: Ratione maxime nulla.
            created_at: "1973-06-07T06:01:32Z"
            id: f0d9b791-a562-4ab9-a803-f4b481c50716
            llm_model: Voluptatem nulla laboriosam qui eos nam rem.
            llm_provider: Aut hic doloribus non.
            metrics:
                aggregates:
                    Laboriosam dolorem.:
                        avg: 0.5605439800138888
                        count: 2098515544
                        max: 0.7862947234767059
                        min: 0.9360853593719282
                        sum: 0.7603072148785976
                period_current_label: Dolores cumque rerum hic accusantium qui repellat.
                period_previous_label: Rerum harum.
                time_series:
                    Ea inventore voluptatum.:
                        anomalies:
                            - period_label: Quibusdam nisi ut ut.
                              reason: Non quo et possimus.
                              value: 0.2705927078818669
                            - period_label: Quibusdam nisi ut ut.
                              reason: Non quo et possimus.
                              value: 0.2705927078818669
                            - period_label: Quibusdam nisi ut ut.
                              reason: Non quo et possimus.
                              value: 0.2705927078818669
                        change: 0.20166130308069102
                        change_percentage: 0.7671929932026079
                        current_period: 0.6419130654630286
                        moving_average: 0.7045118993398898
                        periods:
                            - label: Suscipit eum magni cupiditate.
                              value: 0.9598789242890686
                            - label: Suscipit eum magni cupiditate.
                              value: 0.9598789242890686
                            - label: Suscipit eum magni cupiditate.
                              value: 0.9598789242890686
                        previous_period: 0.18823648238661542
                        trend: Adipisci consectetur possimus culpa ipsam.
                        trend_summary:
                            direction: Maiores et aut.
                            periods_used: 326161229
                            slope: 0.8004050101397506
                            summary: Id aut veniam dignissimos fugit.
                    Officia quod.:
                        anomalies:
                            - period_label: Quibusdam nisi ut ut.
                              reason: Non quo et possimus.
                              value: 0.2705927078818669
                            - period_label: Quibusdam nisi ut ut.
                              reason: Non quo et possimus.
                              value: 0.2705927078818669
                            - period_label: Quibusdam nisi ut ut.
                              reason: Non quo et possimus.
                              value: 0.2705927078818669
                        change: 0.20166130308069102
                        change_percentage: 0.7671929932026079
                        current_period: 0.6419130654630286
                        moving_average: 0.7045118993398898
                        periods:
                            - label: Suscipit eum magni cupiditate.
                              value: 0.9598789242890686
                            - label: Suscipit eum magni cupiditate.
                              value: 0.9598789242890686
                            - label: Suscipit eum magni cupiditate.
                              value: 0.9598789242890686
                        previous_period: 0.18823648238661542
                        trend: Adipisci consectetur possimus culpa ipsam.
                        trend_summary:
                            direction: Maiores et aut.
                            periods_used: 326161229
                            slope: 0.8004050101397506
                            summary: Id aut veniam dignissimos fugit.
                    Optio facere repellat dolorem.:
                        anomalies:
                            - period_label: Quibusdam nisi ut ut.
                              reason: Non quo et possimus.
                              value: 0.2705927078818669
                            - period_label: Quibusdam nisi ut ut.
                              reason: Non quo et possimus.
                              value: 0.2705927078818669
                            - period_label: Quibusdam nisi ut ut.
                              reason: Non quo et possimus.
                              value: 0.2705927078818669
                        change: 0.20166130308069102
                        change_percentage: 0.7671929932026079
                        current_period: 0.6419130654630286
                        moving_average: 0.7045118993398898
                        periods:
                            - label: Suscipit eum magni cupiditate.
                              value: 0.9598789242890686
                            - label: Suscipit eum magni cupiditate.
                              value: 0.9598789242890686
                            - label: Suscipit eum magni cupiditate.
                              value: 0.9598789242890686
                        previous_period: 0.18823648238661542
                        trend: Adipisci consectetur possimus culpa ipsam.
                        trend_summary:
                            direction: Maiores et aut.
                            periods_used: 326161229
                            slope: 0.8004050101397506
                            summary: Id aut veniam dignissimos fugit.
                top_categories:
                    Omnis qui velit molestiae.:
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
                        - category: Est ut quia.
                          percentage: 0.4370289696293134
                          value: 0.07511843322601428
            narrative:
                drivers:
                    - Est non et nihil officiis.
                    - Eum et vitae distinctio dolorem.
                    - Dolorum porro voluptas laudantium.
                    - Veritatis minus reiciendis laudantium.
                headline: Quae qui similique.
                limitations:
                    - Eius fugit a doloremque.
                    - Aut eveniet non sed velit.
                recommendations:
                    - Tempora et natus.
                    - Dolor quibusdam.
                    - Natus perspiciatis aperiam.
                    - Earum est explicabo.
                takeaways:
                    - Est temporibus vel maxime omnis iste.
                    - Non nesciunt incidunt illo molestias.
            saved_query_id: a0b8c451-ebd6-49a0-96a5-bde413f24a5d
            sql: Quia velit consequatur officiis.
        required:
            - id
            - sql
            - narrative
            - metrics
            - created_at
            - llm_model
            - llm_provider
...
    ValidationError:
        title: ValidationError
        type: object
        properties:
            code:
                type: string
                example: Dolores voluptates eius minus.
            message:
                type: string
                example: Reiciendis sit placeat vero.
            name:
                type: string
                example: Incidunt quisquam vero consequatur.
        example:
            code: Quidem illum adipisci omnis sed molestiae.
            message: Similique ut suscipit reiciendis dolor.
            name: Animi odit sed dolore perferendis.
        required:
            - name
            - message