# api/gen/http/reports/server/encode_decode.go
// Code generated by goa v3.24.1, DO NOT EDIT.
//
// reports HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/pgquerynarrative/pgquerynarrative/api/design

package server

import (
	"context"
	"errors"
	"io"
	"net/http"
	"strconv"

	reports "github.com/pgquerynarrative/pgquerynarrative/api/gen/reports"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeGenerateResponse returns an encoder for responses returned by the
// reports generate endpoint.
func EncodeGenerateResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*reports.Report)
		enc := encoder(ctx, w)
		body := NewGenerateResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGenerateRequest returns a decoder for requests sent to the reports
// generate endpoint.
func DecodeGenerateRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*reports.GenerateReportPayload, error) {
	return func(r *http.Request) (*reports.GenerateReportPayload, error) {
		var (
			body GenerateRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateGenerateRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewGenerateReportPayload(&body)

		return payload, nil
	}
}

// EncodeGenerateError returns an encoder for errors returned by the generate
// reports endpoint.
func EncodeGenerateError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "llm_error":
			var res *reports.LLMError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGenerateLlmErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return enc.Encode(body)
		case "validation_error":
			var res *reports.ValidationError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGenerateValidationErrorResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeGetResponse returns an encoder for responses returned by the reports
// get endpoint.
func EncodeGetResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*reports.Report)
		enc := encoder(ctx, w)
		body := NewGetResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetRequest returns a decoder for requests sent to the reports get
// endpoint.
func DecodeGetRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*reports.GetPayload, error) {
	return func(r *http.Request) (*reports.GetPayload, error) {
		var (
			id  string
			err error

			params = mux.Vars(r)
		)
		id = params["id"]
		err = goa.MergeErrors(err, goa.ValidateFormat("id", id, goa.FormatUUID))
		if err != nil {
			return nil, err
		}
		payload := NewGetPayload(id)

		return payload, nil
	}
}

// EncodeGetError returns an encoder for errors returned by the get reports
// endpoint.
func EncodeGetError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "not_found":
			var res *reports.NotFoundError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewGetNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeListResponse returns an encoder for responses returned by the reports
// list endpoint.
func EncodeListResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*reports.ReportList)
		enc := encoder(ctx, w)
		body := NewListResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeListRequest returns a decoder for requests sent to the reports list
// endpoint.
func DecodeListRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*reports.ListPayload, error) {
	return func(r *http.Request) (*reports.ListPayload, error) {
		var (
			savedQueryID *string
			limit        int32
			offset       int32
			err          error
		)
		qp := r.URL.Query()
		savedQueryIDRaw := qp.Get("saved_query_id")
		if savedQueryIDRaw != "" {
			savedQueryID = &savedQueryIDRaw
		}
		if savedQueryID != nil {
			err = goa.MergeErrors(err, goa.ValidateFormat("saved_query_id", *savedQueryID, goa.FormatUUID))
		}
		{
			limitRaw := qp.Get("limit")
			if limitRaw == "" {
				limit = 50
			} else {
				v, err2 := strconv.ParseInt(limitRaw, 10, 32)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("limit", limitRaw, "integer"))
				}
				limit = int32(v)
			}
		}
		if limit < 1 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("limit", limit, 1, true))
		}
		if limit > 100 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("limit", limit, 100, false))
		}
		{
			offsetRaw := qp.Get("offset")
			if offsetRaw != "" {
				v, err2 := strconv.ParseInt(offsetRaw, 10, 32)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("offset", offsetRaw, "integer"))
				}
				offset = int32(v)
			}
		}
		if offset < 0 {
			err = goa.MergeErrors(err, goa.InvalidRangeError("offset", offset, 0, true))
		}
		if err != nil {
			return nil, err
		}
		payload := NewListPayload(savedQueryID, limit, offset)

		return payload, nil
	}
}

// marshalReportsNarrativeContentToNarrativeContentResponseBody builds a value
// of type *NarrativeContentResponseBody from a value of type
// *reports.NarrativeContent.
func marshalReportsNarrativeContentToNarrativeContentResponseBody(v *reports.NarrativeContent) *NarrativeContentResponseBody {
	res := &NarrativeContentResponseBody{
		Headline: v.Headline,
	}
	if v.Takeaways != nil {
		res.Takeaways = make([]string, len(v.Takeaways))
		for i, val := range v.Takeaways {
			res.Takeaways[i] = val
		}
	} else {
		res.Takeaways = []string{}
	}
	if v.Drivers != nil {
		res.Drivers = make([]string, len(v.Drivers))
		for i, val := range v.Drivers {
			res.Drivers[i] = val
		}
	}
	if v.Limitations != nil {
		res.Limitations = make([]string, len(v.Limitations))
		for i, val := range v.Limitations {
			res.Limitations[i] = val
		}
	}
	if v.Recommendations != nil {
		res.Recommendations = make([]string, len(v.Recommendations))
		for i, val := range v.Recommendations {
			res.Recommendations[i] = val
		}
	}

	return res
}

// marshalReportsMetricsDataToMetricsDataResponseBody builds a value of type
// *MetricsDataResponseBody from a value of type *reports.MetricsData.
func marshalReportsMetricsDataToMetricsDataResponseBody(v *reports.MetricsData) *MetricsDataResponseBody {
	res := &MetricsDataResponseBody{
		PeriodCurrentLabel:  v.PeriodCurrentLabel,
		PeriodPreviousLabel: v.PeriodPreviousLabel,
	}
	if v.Aggregates != nil {
		res.Aggregates = make(map[string]*AggregateDataResponseBody, len(v.Aggregates))
		for key, val := range v.Aggregates {
			tk := key
			if val == nil {
				res.Aggregates[tk] = nil
				continue
			}
			res.Aggregates[tk] = marshalReportsAggregateDataToAggregateDataResponseBody(val)
		}
	}
	if v.TopCategories != nil {
		res.TopCategories = make(map[string][]*TopCategoryDataResponseBody, len(v.TopCategories))
		for key, val := range v.TopCategories {
			tk := key
			tv := make([]*TopCategoryDataResponseBody, len(val))
			for i, val := range val {
				if val == nil {
					tv[i] = nil
					continue
				}
				tv[i] = marshalReportsTopCategoryDataToTopCategoryDataResponseBody(val)
			}
			res.TopCategories[tk] = tv
		}
	}
	if v.TimeSeries != nil {
		res.TimeSeries = make(map[string]*TimeSeriesDataResponseBody, len(v.TimeSeries))
		for key, val := range v.TimeSeries {
			tk := key
			if val == nil {
				res.TimeSeries[tk] = nil
				continue
			}
			res.TimeSeries[tk] = marshalReportsTimeSeriesDataToTimeSeriesDataResponseBody(val)
		}
	}

	return res
}

// marshalReportsAggregateDataToAggregateDataResponseBody builds a value of
// type *AggregateDataResponseBody from a value of type *reports.AggregateData.
func marshalReportsAggregateDataToAggregateDataResponseBody(v *reports.AggregateData) *AggregateDataResponseBody {
	if v == nil {
		return nil
	}
	res := &AggregateDataResponseBody{
		Sum:   v.Sum,
		Avg:   v.Avg,
		Min:   v.Min,
		Max:   v.Max,
		Count: v.Count,
	}

	return res
}

// marshalReportsTopCategoryDataToTopCategoryDataResponseBody builds a value of
// type *TopCategoryDataResponseBody from a value of type
// *reports.TopCategoryData.
func marshalReportsTopCategoryDataToTopCategoryDataResponseBody(v *reports.TopCategoryData) *TopCategoryDataResponseBody {
	if v == nil {
		return nil
	}
	res := &TopCategoryDataResponseBody{
		Category:   v.Category,
		Value:      v.Value,
		Percentage: v.Percentage,
	}

	return res
}

// marshalReportsTimeSeriesDataToTimeSeriesDataResponseBody builds a value of
// type *TimeSeriesDataResponseBody from a value of type
// *reports.TimeSeriesData.
func marshalReportsTimeSeriesDataToTimeSeriesDataResponseBody(v *reports.TimeSeriesData) *TimeSeriesDataResponseBody {
	if v == nil {
		return nil
	}
	res := &TimeSeriesDataResponseBody{
		CurrentPeriod:    v.CurrentPeriod,
		PreviousPeriod:   v.PreviousPeriod,
		Change:           v.Change,
		ChangePercentage: v.ChangePercentage,
		Trend:            v.Trend,
		MovingAverage:    v.MovingAverage,
	}
	if v.Periods != nil {
		res.Periods = make([]*PeriodPointDataResponseBody, len(v.Periods))
		for i, val := range v.Periods {
			if val == nil {
				res.Periods[i] = nil
				continue
			}
			res.Periods[i] = marshalReportsPeriodPointDataToPeriodPointDataResponseBody(val)
		}
	}
	if v.Anomalies != nil {
		res.Anomalies = make([]*AnomalyPointDataResponseBody, len(v.Anomalies))
		for i, val := range v.Anomalies {
			if val == nil {
				res.Anomalies[i] = nil
				continue
			}
			res.Anomalies[i] = marshalReportsAnomalyPointDataToAnomalyPointDataResponseBody(val)
		}
	}
	if v.TrendSummary != nil {
		res.TrendSummary = marshalReportsTrendSummaryDataToTrendSummaryDataResponseBody(v.TrendSummary)
	}

	return res
}

// marshalReportsPeriodPointDataToPeriodPointDataResponseBody builds a value of
// type *PeriodPointDataResponseBody from a value of type
// *reports.PeriodPointData.
func marshalReportsPeriodPointDataToPeriodPointDataResponseBody(v *reports.PeriodPointData) *PeriodPointDataResponseBody {
	if v == nil {
		return nil
	}
	res := &PeriodPointDataResponseBody{
		Label: v.Label,
		Value: v.Value,
	}

	return res
}

// marshalReportsAnomalyPointDataToAnomalyPointDataResponseBody builds a value
// of type *AnomalyPointDataResponseBody from a value of type
// *reports.AnomalyPointData.
func marshalReportsAnomalyPointDataToAnomalyPointDataResponseBody(v *reports.AnomalyPointData) *AnomalyPointDataResponseBody {
	if v == nil {
		return nil
	}
	res := &AnomalyPointDataResponseBody{
		PeriodLabel: v.PeriodLabel,
		Value:       v.Value,
		Reason:      v.Reason,
	}

	return res
}

// marshalReportsTrendSummaryDataToTrendSummaryDataResponseBody builds a value
// of type *TrendSummaryDataResponseBody from a value of type
// *reports.TrendSummaryData.
func marshalReportsTrendSummaryDataToTrendSummaryDataResponseBody(v *reports.TrendSummaryData) *TrendSummaryDataResponseBody {
	if v == nil {
		return nil
	}
	res := &TrendSummaryDataResponseBody{
		Direction:   v.Direction,
		Slope:       v.Slope,
		PeriodsUsed: v.PeriodsUsed,
		Summary:     v.Summary,
	}

	return res
}

// marshalReportsChartSuggestionToChartSuggestionResponseBody builds a value of
// type *ChartSuggestionResponseBody from a value of type
// *reports.ChartSuggestion.
func marshalReportsChartSuggestionToChartSuggestionResponseBody(v *reports.ChartSuggestion) *ChartSuggestionResponseBody {
	if v == nil {
		return nil
	}
	res := &ChartSuggestionResponseBody{
		ChartType: v.ChartType,
		Label:     v.Label,
		Reason:    v.Reason,
	}

	return res
}

// marshalReportsReportToReportResponseBody builds a value of type
// *ReportResponseBody from a value of type *reports.Report.
func marshalReportsReportToReportResponseBody(v *reports.Report) *ReportResponseBody {
	res := &ReportResponseBody{
		ID:           v.ID,
		SavedQueryID: v.SavedQueryID,
		SQL:          v.SQL,
		CreatedAt:    v.CreatedAt,
		LlmModel:     v.LlmModel,
		LlmProvider:  v.LlmProvider,
	}
	if v.Narrative != nil {
		res.Narrative = marshalReportsNarrativeContentToNarrativeContentResponseBody(v.Narrative)
	}
	if v.Metrics != nil {
		res.Metrics = marshalReportsMetricsDataToMetricsDataResponseBody(v.Metrics)
	}
	if v.ChartSuggestions != nil {
		res.ChartSuggestions = make([]*ChartSuggestionResponseBody, len(v.ChartSuggestions))
		for i, val := range v.ChartSuggestions {
			if val == nil {
				res.ChartSuggestions[i] = nil
				continue
			}
			res.ChartSuggestions[i] = marshalReportsChartSuggestionToChartSuggestionResponseBody(val)
		}
	}

	return res
}